+++ libs/ionic-react/CHANGELOG.md
@@ -9,6 +9,7 @@
 - upgrade `@testing-library/jest-dom` to 5.5.0
 - upgrade `@testing-library/cypress` to 6.0.0
 - upgrade `@testing-library/user-event` to 10.0.1
+- honor `unitTestRunner` flag
 
 # 1.0.2